requirements.txt
python-docx>=1.2.0
Pillow>=11.3.0
moviepy>=2.2.1
ffmpeg-python>=0.2.0
pypandoc>=1.15
pdf2docx>=0.5.8
ebooklib>=0.19.0
numpy>=2.2.6
tqdm>=4.67.1
python-dotenv>=1.1.1
typing-extensions>=4.14.1
